organize tests and divide concern
