+++ onchain_tests/__tests__/workspace.ts
@@ -1,0 +1,27 @@
+import { Workspace } from "near-workspaces-ava";
+
+export const workspace = Workspace.init(async ({ root }) => {
+  const alice = await root.createAccount("alice");
+  const bob = await root.createAccount("bob");
+  const charlie = await root.createAccount("charlie");
+
+  const nft_contract = await root.createAndDeploy(
+    "nft-contract",
+    "../out/main.wasm",
+    {
+      method: "new_default_meta",
+      args: { owner_id: root },
+    }
+  );
+
+  const market_contract = await root.createAndDeploy(
+    "nft-market",
+    "../out/market.wasm",
+    {
+      method: "new",
+      args: { owner_id: root },
+    }
+  );
+
+  return { alice, bob, charlie, nft_contract, market_contract };
+});
